% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_fetched_timestamp}
\alias{add_fetched_timestamp}
\title{Add Fetched Timestamp}
\usage{
add_fetched_timestamp(data)
}
\arguments{
\item{data}{Data frame or tibble}
}
\value{
Data frame with fetched_at column
}
\description{
Adds a fetched_at timestamp to data.
}
\keyword{internal}
