% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{adlibr_build_request}
\alias{adlibr_build_request}
\title{Build Platform Request}
\usage{
adlibr_build_request(platform, operation, params)
}
\arguments{
\item{platform}{Character. Platform name}

\item{operation}{Character. Operation type}

\item{params}{List. Request parameters}
}
\value{
Request object (httr2 request or special platform object)
}
\description{
Builds a request object based on platform configuration and parameters.
}
\keyword{internal}
