% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{adlibr_platforms}
\alias{adlibr_platforms}
\title{Get Supported Platforms}
\usage{
adlibr_platforms()
}
\value{
Character vector of platform names
}
\description{
Returns a character vector of all supported platform names.
}
