% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{build_countries_param}
\alias{build_countries_param}
\title{Build Country Parameter String}
\usage{
build_countries_param(countries)
}
\arguments{
\item{countries}{A character vector of lowercase 2-letter ISO country codes}
}
\value{
A URL-encoded string for the countries parameter
}
\description{
Build Country Parameter String
}
\keyword{internal}
