% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{build_google_advertisers_query}
\alias{build_google_advertisers_query}
\title{Build Google Advertisers Query}
\usage{
build_google_advertisers_query(params, config)
}
\arguments{
\item{params}{List. Query parameters}

\item{config}{List. Platform configuration}
}
\value{
Character. SQL query string
}
\description{
Builds SQL query for Google advertiser information.
}
\keyword{internal}
