% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{check_google_bq_auth}
\alias{check_google_bq_auth}
\title{Check Google BigQuery Authentication}
\usage{
check_google_bq_auth()
}
\value{
Logical. TRUE if authentication is available
}
\description{
Checks if Google BigQuery authentication is properly configured.
}
\keyword{internal}
