% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse_list_column}
\alias{collapse_list_column}
\title{Collapse List Column to String}
\usage{
collapse_list_column(x, sep = ", ", max_items = 5)
}
\arguments{
\item{x}{List or vector}

\item{sep}{Character. Separator string}

\item{max_items}{Integer. Maximum items to include}
}
\value{
Character. Collapsed string
}
\description{
Safely collapses list columns to character strings for display.
}
\keyword{internal}
