% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{configure_meta_auth}
\alias{configure_meta_auth}
\title{Configure Meta Authentication (placeholder)}
\usage{
configure_meta_auth(force = FALSE, use_keyring = TRUE)
}
\description{
Configure Meta Authentication (placeholder)
}
\keyword{internal}
