% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{configure_tiktok_auth}
\alias{configure_tiktok_auth}
\title{Configure TikTok Authentication (placeholder)}
\usage{
configure_tiktok_auth(force = FALSE, use_keyring = TRUE)
}
\description{
Configure TikTok Authentication (placeholder)
}
\keyword{internal}
