% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{convert_google_regions_to_countries}
\alias{convert_google_regions_to_countries}
\title{Convert Google Regions to Country Codes}
\usage{
convert_google_regions_to_countries(regions)
}
\arguments{
\item{regions}{Character vector. Google region names}
}
\value{
Character vector. ISO country codes
}
\description{
Converts Google's region names to ISO country codes where possible.
}
\keyword{internal}
