% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft.R
\name{convert_microsoft_code_to_iso}
\alias{convert_microsoft_code_to_iso}
\title{Convert Microsoft Code to ISO Country Code}
\usage{
convert_microsoft_code_to_iso(ms_code)
}
\arguments{
\item{ms_code}{Character. Microsoft country code}
}
\value{
Character. ISO country code or original if not found
}
\description{
Converts Microsoft's numeric country codes back to ISO codes.
}
\keyword{internal}
