% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_user_agent}
\alias{create_user_agent}
\title{Create User Agent String}
\usage{
create_user_agent()
}
\value{
Character. User-Agent string
}
\description{
Creates a standardized User-Agent string for HTTP requests.
}
\keyword{internal}
