% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\name{determine_apple_date_preset}
\alias{determine_apple_date_preset}
\title{Determine Apple Date Preset}
\usage{
determine_apple_date_preset(date_from, date_to)
}
\arguments{
\item{date_from}{Date or character. Start date}

\item{date_to}{Date or character. End date}
}
\value{
Character. Apple date preset
}
\description{
Converts date range to Apple's preset format.
}
\keyword{internal}
