% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{determine_google_dataset}
\alias{determine_google_dataset}
\title{Determine Google Dataset}
\usage{
determine_google_dataset(params)
}
\arguments{
\item{params}{List. Query parameters}
}
\value{
Character. Dataset identifier
}
\description{
Determines which Google dataset to query based on parameters.
}
\keyword{internal}
