% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{execute_google_bq_request}
\alias{execute_google_bq_request}
\title{Execute Google BigQuery Request}
\usage{
execute_google_bq_request(req_config)
}
\arguments{
\item{req_config}{List. Request configuration from build_google_bq_request}
}
\value{
tibble. Query results
}
\description{
Executes BigQuery request and returns results.
}
\keyword{internal}
