% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{extract_linkedin_targeting}
\alias{extract_linkedin_targeting}
\title{Extract LinkedIn Targeting Information}
\usage{
extract_linkedin_targeting(targeting_data, facet_names)
}
\arguments{
\item{targeting_data}{List. LinkedIn targeting data}

\item{facet_names}{Character vector. Facet names to extract}
}
\value{
Character vector. Extracted targeting values
}
\description{
Extracts targeting information from LinkedIn ad targeting data.
}
\keyword{internal}
