% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\name{get_apple_countries}
\alias{get_apple_countries}
\title{Get Apple Countries/Regions}
\usage{
get_apple_countries()
}
\value{
tibble. Available countries/regions
}
\description{
Retrieves list of available countries/regions from Apple API.
}
\keyword{internal}
