% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{get_delay_value}
\alias{get_delay_value}
\title{Get Adaptive Delay Based on App Usage}
\usage{
get_delay_value(call_pct = 0, cpu_pct = 0, time_pct = 0)
}
\arguments{
\item{call_pct}{Numeric. Percentage of allowed call count used.}

\item{cpu_pct}{Numeric. Percentage of allowed CPU time used.}

\item{time_pct}{Numeric. Percentage of allowed total time used.}
}
\value{
A numeric value indicating the number of seconds to sleep.
}
\description{
Calculates the appropriate sleep delay based on Meta API usage percentages (calls, CPU time, total time).
Inspired by the \href{https://github.com/Wesleyan-Media-Project/fb_ads_import/blob/main/race2022_utils.R}{Wesleyan Media Project's utilities}.
}
\keyword{internal}
