% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{\%||\%}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value}
}
\value{
The first non-NULL value
}
\description{
Returns the first non-NULL value from left to right.
}
\keyword{internal}
