% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R, R/errors.R
\name{handle_response_errors}
\alias{handle_response_errors}
\title{Handle Response Errors}
\usage{
handle_response_errors(response, platform, endpoint)

handle_response_errors(response, platform, endpoint)
}
\arguments{
\item{response}{httr2 response object}

\item{platform}{Character. The platform name}

\item{endpoint}{Character. The endpoint URL}

\item{operation}{Character. Operation type}
}
\value{
Invisible NULL if no errors

The response object if no errors, otherwise throws an error
}
\description{
Checks response for errors and throws appropriate custom exceptions

Checks an HTTP response for errors and throws appropriate adlibr error conditions.
}
\keyword{internal}
