% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{li_auth}
\alias{li_auth}
\title{Authenticate with LinkedIn (3-Legged OAuth)}
\usage{
li_auth(scope = "openid profile", cache = TRUE, new_user = FALSE)
}
\arguments{
\item{scope}{A space-delimited string of permissions your application is
requesting. Per new OpenID Connect docs, this defaults to "openid profile".}

\item{cache}{Logical or a string. \code{TRUE} caches the token in a file named
\code{.httr-oauth} in the current working directory. Caching is highly recommended.}

\item{new_user}{Logical. Set to \code{TRUE} to force a new authentication, even if
a cached token exists.}
}
\value{
A \code{Token2.0} object, which is automatically cached.
}
\description{
This is the \strong{second step}. It initiates the 3-legged OAuth 2.0 flow.
It will open your web browser, ask you to log in to LinkedIn and authorize the
application scopes. After authorization, it automatically fetches and caches
an access token for future use.
}
\examples{
\dontrun{
# After configuring, run this to authenticate. Your browser will open.
li_token <- li_auth()
}
}
