% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{manual_linkedin_oauth}
\alias{manual_linkedin_oauth}
\title{Manual LinkedIn OAuth Flow}
\usage{
manual_linkedin_oauth(client_id, client_secret)
}
\arguments{
\item{client_id}{Character. LinkedIn client ID}

\item{client_secret}{Character. LinkedIn client secret}
}
\value{
httr Token object
}
\description{
Performs OAuth flow manually for non-interactive environments
}
\keyword{internal}
