% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon.R
\name{normalize_amazon_enhanced}
\alias{normalize_amazon_enhanced}
\title{Normalize Amazon Data to Common Schema}
\usage{
normalize_amazon_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Amazon API response}
}
\value{
tibble. Normalized data
}
\description{
Converts Amazon EU Ad Library data to adlibr normalized schema.
}
\keyword{internal}
