% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_apple}
\alias{normalize_apple}
\title{Normalize Apple Data}
\usage{
normalize_apple(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Apple API response}
}
\value{
tibble. Normalized data
}
\description{
Normalize Apple Data
}
\keyword{internal}
