% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booking.R
\name{normalize_booking_enhanced}
\alias{normalize_booking_enhanced}
\title{Normalize Booking Data to Common Schema}
\usage{
normalize_booking_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Booking API response}
}
\value{
tibble. Normalized data
}
\description{
Converts Booking.com data to adlibr normalized schema.
Note: Booking.com may have limited ad transparency data.
}
\keyword{internal}
