% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_country_codes}
\alias{normalize_country_codes}
\title{Normalize Country Codes}
\usage{
normalize_country_codes(countries)
}
\arguments{
\item{countries}{Character vector. Country codes to normalize}
}
\value{
Character vector. Normalized country codes
}
\description{
Normalizes country codes to uppercase ISO 3166-1 alpha-2 format.
}
\keyword{internal}
