% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_generic}
\alias{normalize_generic}
\title{Generic Normalization Function}
\usage{
normalize_generic(raw_data, platform)
}
\arguments{
\item{raw_data}{Raw platform data}

\item{platform}{Platform name}
}
\value{
tibble. Normalized data with minimal mapping
}
\description{
Generic normalization for platforms without specific implementations.
}
\keyword{internal}
