% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_bq.R
\name{normalize_google_bq_enhanced}
\alias{normalize_google_bq_enhanced}
\title{Normalize Google BigQuery Data to Common Schema}
\usage{
normalize_google_bq_enhanced(raw_data)
}
\arguments{
\item{raw_data}{tibble. Raw BigQuery results}
}
\value{
tibble. Normalized data
}
\description{
Converts Google BigQuery political ads data to adlibr normalized schema.
}
\keyword{internal}
