% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{normalize_linkedin_enhanced}
\alias{normalize_linkedin_enhanced}
\title{Normalize LinkedIn Data to Common Schema}
\usage{
normalize_linkedin_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw LinkedIn API response}
}
\value{
tibble. Normalized data
}
\description{
Converts LinkedIn API data to adlibr normalized schema.
}
\keyword{internal}
