% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_meta}
\alias{normalize_meta}
\title{Normalize Meta (Facebook/Instagram) Data}
\usage{
normalize_meta(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Meta API response}
}
\value{
tibble. Normalized data
}
\description{
Normalize Meta (Facebook/Instagram) Data
}
\keyword{internal}
