% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{normalize_meta_enhanced}
\alias{normalize_meta_enhanced}
\title{Normalize Meta Data to Common Schema}
\usage{
normalize_meta_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Meta API response}
}
\value{
tibble. Normalized data
}
\description{
Enhanced Meta data normalization with improved field mapping.
}
\keyword{internal}
