% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_microsoft}
\alias{normalize_microsoft}
\title{Normalize Microsoft (Bing) Data}
\usage{
normalize_microsoft(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Microsoft API response}
}
\value{
tibble. Normalized data
}
\description{
Normalize Microsoft (Bing) Data
}
\keyword{internal}
