% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinterest.R
\name{normalize_pinterest_enhanced}
\alias{normalize_pinterest_enhanced}
\title{Normalize Pinterest Data to Common Schema}
\usage{
normalize_pinterest_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Pinterest API response}
}
\value{
tibble. Normalized data
}
\description{
Converts Pinterest Ad Library data to adlibr normalized schema.
}
\keyword{internal}
