% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapchat.R
\name{normalize_snapchat_enhanced}
\alias{normalize_snapchat_enhanced}
\title{Normalize Snapchat Data to Common Schema}
\usage{
normalize_snapchat_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw Snapchat API response}
}
\value{
tibble. Normalized data
}
\description{
Converts Snapchat Political Ads data to adlibr normalized schema.
}
\keyword{internal}
