% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_tiktok}
\alias{normalize_tiktok}
\title{Normalize TikTok Data}
\usage{
normalize_tiktok(raw_data)
}
\arguments{
\item{raw_data}{List. Raw TikTok API response}
}
\value{
tibble. Normalized data
}
\description{
Normalize TikTok Data
}
\keyword{internal}
