% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiktok.R
\name{normalize_tiktok_enhanced}
\alias{normalize_tiktok_enhanced}
\title{Normalize TikTok Data to Common Schema}
\usage{
normalize_tiktok_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw TikTok API response}
}
\value{
tibble. Normalized data
}
\description{
Converts TikTok Commercial Content Library data to adlibr normalized schema.
}
\keyword{internal}
