% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{normalize_x_enhanced}
\alias{normalize_x_enhanced}
\title{Normalize X Data to Common Schema}
\usage{
normalize_x_enhanced(raw_data)
}
\arguments{
\item{raw_data}{List. Raw X API response}
}
\value{
tibble. Normalized data
}
\description{
Converts X Ads Transparency data to adlibr normalized schema.
}
\keyword{internal}
