% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazon.R
\name{parse_amazon_response}
\alias{parse_amazon_response}
\title{Parse Amazon Response}
\usage{
parse_amazon_response(response, operation)
}
\arguments{
\item{response}{httr2 response object}

\item{operation}{Character. Operation type}
}
\value{
List. Parsed response data
}
\description{
Parses Amazon API response and extracts relevant data.
}
\keyword{internal}
