% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\name{parse_apple_response}
\alias{parse_apple_response}
\title{Parse Apple Response}
\usage{
parse_apple_response(response, operation)
}
\arguments{
\item{response}{httr2 response object}

\item{operation}{Character. Operation type}
}
\value{
List. Parsed response data
}
\description{
Parses Apple API response and extracts relevant data.
}
\keyword{internal}
