% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{parse_meta_response}
\alias{parse_meta_response}
\title{Parse Meta Response}
\usage{
parse_meta_response(response)
}
\arguments{
\item{response}{httr2 response object}
}
\value{
List. Parsed response data
}
\description{
Parses Meta Graph API response and extracts ads data.
}
\keyword{internal}
