% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft.R
\name{parse_microsoft_response}
\alias{parse_microsoft_response}
\title{Parse Microsoft Response}
\usage{
parse_microsoft_response(response, operation)
}
\arguments{
\item{response}{httr2 response object}

\item{operation}{Character. Operation type}
}
\value{
List. Parsed response data
}
\description{
Parses Microsoft API response and extracts relevant data.
}
\keyword{internal}
