% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinterest.R
\name{parse_pinterest_response}
\alias{parse_pinterest_response}
\title{Parse Pinterest Response}
\usage{
parse_pinterest_response(response, operation)
}
\arguments{
\item{response}{httr2 response object}

\item{operation}{Character. Operation type}
}
\value{
List. Parsed response data
}
\description{
Parses Pinterest API response and extracts relevant data.
}
\keyword{internal}
