% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.R
\name{parse_x_response}
\alias{parse_x_response}
\title{Parse X Response}
\usage{
parse_x_response(response, operation)
}
\arguments{
\item{response}{httr2 response object}

\item{operation}{Character. Operation type}
}
\value{
List. Parsed response data
}
\description{
Parses X API response and extracts relevant data.
}
\keyword{internal}
