% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apple.R
\name{search_apple_advertisers}
\alias{search_apple_advertisers}
\title{Search Apple Advertisers/Developers}
\usage{
search_apple_advertisers(search_term, limit = 100)
}
\arguments{
\item{search_term}{Character. Search term for developer/app names}

\item{limit}{Integer. Maximum number of results}
}
\value{
tibble. Developer/advertiser information
}
\description{
Search for app developers on Apple platform.
}
\keyword{internal}
