% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoft.R
\name{search_microsoft_advertisers}
\alias{search_microsoft_advertisers}
\title{Search Microsoft Advertisers}
\usage{
search_microsoft_advertisers(search_term, limit = 100)
}
\arguments{
\item{search_term}{Character. Search term for advertiser names}

\item{limit}{Integer. Maximum number of results}
}
\value{
tibble. Advertiser information
}
\description{
Search for advertisers on Microsoft platform.
}
\keyword{internal}
