% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{store_credential}
\alias{store_credential}
\title{Store Credential}
\usage{
store_credential(key, value, use_keyring = TRUE)
}
\arguments{
\item{key}{Character. Credential key}

\item{value}{Character. Credential value}

\item{use_keyring}{Logical. Whether to use keyring}
}
\value{
Invisible TRUE
}
\description{
Stores a credential in keyring or environment file
}
\keyword{internal}
