% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{update_pagination_token}
\alias{update_pagination_token}
\title{Update Pagination Token in Request}
\usage{
update_pagination_token(req, platform, next_token)
}
\arguments{
\item{req}{httr2 request object}

\item{platform}{Character. Platform name}

\item{next_token}{Character. Next page token}
}
\value{
Modified httr2 request object
}
\description{
Updates request with next page token based on platform pagination strategy.
}
\keyword{internal}
