% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{build_daterange_param}
\alias{build_daterange_param}
\title{Build Date Range Parameter String}
\usage{
build_daterange_param(start_date, end_date)
}
\arguments{
\item{start_date}{The start date}

\item{end_date}{The end date}
}
\value{
A URL-encoded string for the dateRange parameter
}
\description{
Build Date Range Parameter String
}
\keyword{internal}
